// lib/api/generated/models/exerciseRecordStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type ExerciseRecordStatus = typeof ExerciseRecordStatus[keyof typeof ExerciseRecordStatus];


export const ExerciseRecordStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  needs_edit: 'needs_edit',
} as const;