// lib/api/generated/models/libraryExercise.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { ExerciseFamily } from './exerciseFamily';
import type { PctOf } from './pctOf';

export interface LibraryExercise {
  default_pct_of?: null | PctOf;
  family?: ExerciseFamily;
  id: string;
  name: string;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
}